implemented drives endpoint

+++ migrations/002.undo.create-drives-table.sql
@@ -1,1 +1,2 @@
-drop table drives;
+drop table drives if exists;
+drop type cloud if exists;
